refactor(logger): evict misfiled host concerns from @logger
@logger had become a junk drawer mixing four unrelated concerns:
trace surface, sink primitives, TeXRA transcript persistence, Supabase
telemetry, agent-domain types. This was the blocker the
agent-trace-sdk-surface proposal flagged for SDK exposure.

Phase A of the cleanup moves the misfiled pieces to where they belong:

- src/logger/TaskState.ts             → src/agent/core/TaskState.ts
- src/logger/AgentUsageReporter.ts    → src/agent/runtime/AgentUsageReporter.ts
- src/logger/streamUtils.ts           → src/agent/runtime/streamTab.ts
- src/logger/UsageLogService.ts       → src/telemetry/UsageLogService.ts
- src/logger/UsageLogTypes.ts         → src/telemetry/UsageLogTypes.ts

Adds @telemetry path alias. Updates 15 consumer files. Drops the moved
re-exports from @logger/index.

Pure relocation — no behavior change. Remaining in @logger: the
TexraTrace extension over @agent/trace, sink primitives, and the
transcript store/recorder (the latter moves out in Phase B).
