+++ packages/desktop/tsconfig.paths.json
@@ -14,6 +14,8 @@
       "@utils/*": ["src/utils/*"],
       "@logger": ["src/logger"],
       "@logger/*": ["src/logger/*"],
+      "@telemetry": ["src/telemetry"],
+      "@telemetry/*": ["src/telemetry/*"],
       "@latex": ["src/latex"],
       "@latex/*": ["src/latex/*"],
       "@commands/*": ["packages/extension/src/commands/*"],
