+++ packages/extension/src/progressView/persistence/mementoMigration.ts
@@ -4,9 +4,9 @@
  */
 
 import type { AgentTrace } from '@agent/trace';
+import { TaskStateSchema, type TaskState } from '@agent/core/TaskState';
 import { WorkspaceStateKey } from '@common/state';
 import { normalizeRunId } from '@common/constants/runIds';
-import { TaskStateSchema, type TaskState } from '@logger/TaskState';
 import {
   StorageRecordSchema,
   type StreamTabId,
