+++ packages/extension/tsconfig.json
@@ -21,6 +21,8 @@
       "@utils/*": ["../../src/utils/*"],
       "@logger": ["../../src/logger"],
       "@logger/*": ["../../src/logger/*"],
+      "@telemetry": ["../../src/telemetry"],
+      "@telemetry/*": ["../../src/telemetry/*"],
       "@skills/*": ["../../src/skills/*"],
       "@latex": ["../../src/latex"],
       "@latex/*": ["../../src/latex/*"],
