+++ src/logger/index.ts
@@ -1,13 +1,5 @@
-export { AgentUsageReporter } from './AgentUsageReporter';
 export * as logUtils from './logUtils';
 export { redactSecrets, type LogRedactionOptions } from './redaction';
-export { getStreamTabId } from './streamUtils';
-export {
-  isToolUseTaskState,
-  isWorkflowTaskState,
-  type TaskState,
-} from './TaskState';
-export { UsageLogService } from './UsageLogService';
 export {
   attachTranscriptRecorder,
   type TranscriptRecorderHandle,
