+++ src/telemetry/UsageLogService.ts
